feat: 组件编译输出调整，demo调整 #126955312 # Reviewed, transaction id: 60091
